got the basics up

+++ WhenPushComesToShove/Assets/3-Scripts/LevelSystems/MinigameData.cs
@@ -5,11 +5,11 @@ using System;
 
 public class MinigameData : MonoBehaviour
 {
-    public int[] scores = new int[4];
-    public static Action<int, int> onScoreAdded;
+    public float[] scores = new float[4];
+    public static Action<int, float> onScoreAdded;
     private bool canUpdateScore = true;
 
-    public void AddScoreForTeam(int teamIndex, int scoreToAdd)
+    public void AddScoreForTeam(int teamIndex, float scoreToAdd)
     {
         if (!canUpdateScore)
         {
@@ -48,7 +48,7 @@ public class MinigameData : MonoBehaviour
         }
 
         //Cleanup just in case
-        scores = new int[4];
+        scores = new float[4];
     }
 
     public int GetHighestScoreIndex()
